requirements.txt
click
molgenis-py-client
